I noticed repeated warnings in the output: ' VisibleDeprecationWarning: using a non-integer number instead of an integer will result in an error in the future '
I traced it down to the following lines (251, 592, 601, 602) in experiment_class.py
and line 25 in plotter.py

For line 251 the problem is np.rint()

initial_timestep = np.rint(self.reaction_profile.shape[0] * experiment.EXPERIMENTAL_KEQ_SAMPLING_RANGE)

If we look at the usuage of numpy.rint()
http://docs.scipy.org/doc/numpy-dev/reference/generated/numpy.rint.html
we can see that it still returns a non integer
I replaced this function call with numpy.int16() as we don't need more precision and it should be faster, which is important as this function is called every integration step.

For lines 25, 592, 601, 602 I converted the floating point division to floor division since the warning informs us that only integer indices will be acceptable in the future.